+++ CHANGELOG.md
@@ -28,6 +28,7 @@ _What's new_
 * Convert credential for opsgenie to connection. ([#477](https://github.com/turbot/pipe-fittings/issues/477)).
 * Convert credential for pagerduty to connection. ([#478](https://github.com/turbot/pipe-fittings/issues/478)).
 * Convert credential for sendgrid to connection. ([#479](https://github.com/turbot/pipe-fittings/issues/479)).
+* Convert credential for servicenow to connection. ([#480](https://github.com/turbot/pipe-fittings/issues/480)).
 
 _Bug fixes_